+++ Pal/src/host/Linux-SGX/sgx_enclave.c
@@ -431,7 +431,7 @@ static int sgx_ocall_sock_recv(void * pms)
     struct sockaddr * addr = ms->ms_addr;
     socklen_t addrlen = ms->ms_addr ? ms->ms_addrlen : 0;
 
-    if (ms->ms_sockfd == PAL_SEC()->mcast_srv)
+    if ((uint32_t)ms->ms_sockfd == PAL_SEC()->mcast_srv)
         addr = NULL;
 
     ret = INLINE_SYSCALL(recvfrom, 6,
@@ -452,7 +452,7 @@ static int sgx_ocall_sock_send(void * pms)
     const struct sockaddr * addr = ms->ms_addr;
     socklen_t addrlen = ms->ms_addr ? ms->ms_addrlen : 0;
 
-    if (ms->ms_sockfd == PAL_SEC()->mcast_srv) {
+    if ((uint32_t)ms->ms_sockfd == PAL_SEC()->mcast_srv) {
         struct sockaddr_in * mcast_addr = __alloca(sizeof(struct sockaddr_in));
         mcast_addr->sin_family = AF_INET;
         inet_pton4(MCAST_GROUP, sizeof(MCAST_GROUP),  &mcast_addr->sin_addr.s_addr);
